+++ src/Command.js
@@ -22,8 +22,11 @@ class Command extends AbstractCommand {
     return new Proxy(this, {
       get (target, name) {
         if (['run', 'instance', 'class'].includes(name)) return target[name]
-        else if (typeof name === 'string' && name.startsWith('get')) return (...args) => target.instance[name].call(null, ...args)
-        else return target.instance[name]
+        else if (typeof name === 'string' && name.startsWith('get')) {
+          return (...args) => {
+            return target.instance[name](...args)
+          }
+        } else return target.instance[name]
       }
     })
   }