+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+3.3.0
+* Bundle size decrease using [size-limit](https://www.npmjs.com/package/size-limit)
+* `HashCommand` updates. Removed method `HashCommand.getStringHash`, added new `HashCommand.getInputHash`
+* Updated all dependencies to the latest versions
+* Test updates
+* Better coverage
+* Docs updated
+* Other fixes
+
 3.2.1
 * Updated API version to `v3`
 * Added new option for FetchCommand `formatting` (enabled by default). This option add header `X-Requested-With: XMLHttpRequest` which is useful for performance increase.
@@ -6,7 +15,7 @@
 3.2.0
 * New endpoint added `account/status`
 * Tests updates
-* Account class added
+* `Account` class added
 * Docs updated
 
 3.1.0